poi_from_centroid/requirements.txt
google-cloud-bigquery
requests
python-dotenv
gunicorn